Add docker-compose
It's nice to be able to use this image for running tests or other
commands in CI before actually using the build script. Some of our repos
have (or could have!) simple docker-compose setups that will improve
parity between CI and local dev. This allows us to re-use this image for
that purpose.