Incorporate ECR workflow by default if context is present
Also add a number of ECR commands, including commands for managing the
lifecycle policy, and a default lifecycle policy.

This adds another standard tag, "$BRANCH-$BUILD", which is important for
implementing lifecycle rules that don't inadvertently expire images that
are deployed to prod. This works by preserving all images that have tags
prefixed with "main-" or "master-".